+++ lmcache/v1/multiprocess/transport/shm.py
@@ -10,12 +10,12 @@ from typing import Any
 import torch
 
 # First Party
-from lmcache.v1.multiprocess.non_gpu_context import (
+from lmcache.v1.multiprocess.protocol import RequestType, get_response_class
+from lmcache.v1.multiprocess.transport.base import (
     NonGpuContext,
     NonGpuContextMetadata,
 )
-from lmcache.v1.multiprocess.protocol import RequestType, get_response_class
-from lmcache.v1.multiprocess.shm_types import ShmSlotDescriptor
+from lmcache.v1.multiprocess.transport.shm_types import ShmSlotDescriptor
 
 
 class NonGpuContextShm(NonGpuContext):
